In out report update included Permanent Shift

+++ hrms/attendance/roster/roster-list.aspx
@@ -48,7 +48,7 @@
             <div class="container-fluid">
                 <div class="row justify-content-center">
                     <div class="col-lg-3 col-sm-12 mb-lg-0 mb-30">
-                        <div class="card">
+                        <div class="card" id="unitSection">
                             <div id="toggleFilter" class="card-header px-20 py-15" style="cursor: pointer;">
                                 <h6 class="d-flex justify-between align-items-center fw-500 w-100">
                                     <span class="d-flex align-items-center" style="font-size:16px; color:black";>
@@ -376,6 +376,27 @@
                 const fromDate = $('#txtStartDate').val();
                 const toDate = $('#txtEndDate').val();
 
+                if (fromDate == null || fromDate == '') {
+                    Swal.fire({
+                        title: 'Warning!',
+                        text: 'Please Select Start Date',
+                        icon: 'warning',
+                        confirmButtonText: 'OK'
+                    })
+                    return
+
+                }
+
+                if (toDate == null || toDate == '') {
+                    Swal.fire({
+                        title: 'Warning!',
+                        text: 'Please Select End Date',
+                        icon: 'warning',
+                        confirmButtonText: 'OK'
+                    })
+                    return
+                }
+
                 sessionStorage.setItem('__startDate__', fromDate);
                 sessionStorage.setItem('__endDate__', toDate);
 
@@ -586,8 +607,14 @@
                         if (response.statusCode === 200) {
                             var responseData = response.data;
                             console.log('Before table Data Bind', responseData);
-
                             bindUnits(responseData);
+                            if (responseData.length >1) {
+                                $('#unitSection').show();
+                            } else {
+                                $('#unitSection').hide();
+                            }
+
+                       
 
                             console.log('after Table Data Bind ', responseData);
                         } else {
